fix id types for schema validation

+++ color-manager/src/main/java/com/my_wall_color/color_manager/ColorController.java
@@ -19,7 +19,7 @@ public class ColorController {
     }
 
     @GetMapping("/{id}")
-    private ResponseEntity<JpaColor> getColorById(@PathVariable long id) {
+    private ResponseEntity<JpaColor> getColorById(@PathVariable int id) {
         return ResponseEntity.of(this.colorRepository.findById(id));
     }
 }
